requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Single-cell analysis
scanpy>=1.9.0
anndata>=0.9.0
scvi-tools>=1.0.0
scvelo>=0.3.0
celltypist>=1.5.0

# Machine learning
scikit-learn>=1.3.0
umap-learn>=0.5.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data processing
h5py>=3.8.0
zarr<3.0.0
scikit-misc>=0.1.4

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Model sharing
huggingface-hub>=0.19.0 